refactor: move assertSnapshot to snapshot.ts
