+++ testing/_diff.ts
@@ -1,6 +1,16 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
+import {
+  bgGreen,
+  bgRed,
+  bold,
+  gray,
+  green,
+  red,
+  white,
+} from "../fmt/colors.ts";
+
 interface FarthestPoint {
   y: number;
   id: number;
@@ -357,4 +367,69 @@ export function diffstr(A: string, B: string) {
   }
 
   return diffResult;
+}
+
+/**
+ * Colors the output of assertion diffs
+ * @param diffType Difference type, either added or removed
+ */
+function createColor(
+  diffType: DiffType,
+  { background = false } = {},
+): (s: string) => string {
+  switch (diffType) {
+    case DiffType.added:
+      return (s: string): string =>
+        background ? bgGreen(white(s)) : green(bold(s));
+    case DiffType.removed:
+      return (s: string): string => background ? bgRed(white(s)) : red(bold(s));
+    default:
+      return white;
+  }
+}
+
+/**
+ * Prefixes `+` or `-` in diff output
+ * @param diffType Difference type, either added or removed
+ */
+function createSign(diffType: DiffType): string {
+  switch (diffType) {
+    case DiffType.added:
+      return "+   ";
+    case DiffType.removed:
+      return "-   ";
+    default:
+      return "    ";
+  }
+}
+
+export function buildMessage(
+  diffResult: ReadonlyArray<DiffResult<string>>,
+  { stringDiff = false } = {},
+): string[] {
+  const messages: string[] = [], diffMessages: string[] = [];
+  messages.push("");
+  messages.push("");
+  messages.push(
+    `    ${gray(bold("[Diff]"))} ${red(bold("Actual"))} / ${
+      green(bold("Expected"))
+    }`,
+  );
+  messages.push("");
+  messages.push("");
+  diffResult.forEach((result: DiffResult<string>): void => {
+    const c = createColor(result.type);
+    const line = result.details?.map((detail) =>
+      detail.type !== DiffType.common
+        ? createColor(detail.type, { background: true })(detail.value)
+        : detail.value
+    ).join("") ?? result.value;
+    diffMessages.push(c(`${
+      createSign(result.type)
+    }${line}`));
+  });
+  messages.push(...(stringDiff ? [diffMessages.join("")] : diffMessages));
+  messages.push("");
+
+  return messages;
 }
