+++ testing/asserts.ts
@@ -2,112 +2,17 @@
 // This module is browser compatible. Do not rely on good formatting of values
 // for AssertionError messages in browsers.
 
-import {
-  bgGreen,
-  bgRed,
-  bold,
-  gray,
-  green,
-  red,
-  stripColor,
-  white,
-} from "../fmt/colors.ts";
-import { fromFileUrl, join, parse } from "../path/mod.ts";
-import { ensureFile, ensureFileSync } from "../fs/mod.ts";
-import { diff, DiffResult, diffstr, DiffType } from "./_diff.ts";
+import { red, stripColor } from "../fmt/colors.ts";
+import { buildMessage, diff, diffstr } from "./_diff.ts";
+import { format } from "./_format.ts";
 
 const CAN_NOT_DISPLAY = "[Cannot display]";
-const SNAPSHOT_DIR = "__snapshots__";
 
 export class AssertionError extends Error {
   override name = "AssertionError";
   constructor(message: string) {
     super(message);
   }
-}
-
-/**
- * Converts the input into a string. Objects, Sets and Maps are sorted so as to
- * make tests less flaky
- * @param v Value to be formatted
- */
-export function _format(v: unknown): string {
-  // deno-lint-ignore no-explicit-any
-  const { Deno } = globalThis as any;
-  return typeof Deno?.inspect === "function"
-    ? Deno.inspect(v, {
-      depth: Infinity,
-      sorted: true,
-      trailingComma: true,
-      compact: false,
-      iterableLimit: Infinity,
-    })
-    : `"${String(v).replace(/(?=["\\])/g, "\\")}"`;
-}
-
-/**
- * Colors the output of assertion diffs
- * @param diffType Difference type, either added or removed
- */
-function createColor(
-  diffType: DiffType,
-  { background = false } = {},
-): (s: string) => string {
-  switch (diffType) {
-    case DiffType.added:
-      return (s: string): string =>
-        background ? bgGreen(white(s)) : green(bold(s));
-    case DiffType.removed:
-      return (s: string): string => background ? bgRed(white(s)) : red(bold(s));
-    default:
-      return white;
-  }
-}
-
-/**
- * Prefixes `+` or `-` in diff output
- * @param diffType Difference type, either added or removed
- */
-function createSign(diffType: DiffType): string {
-  switch (diffType) {
-    case DiffType.added:
-      return "+   ";
-    case DiffType.removed:
-      return "-   ";
-    default:
-      return "    ";
-  }
-}
-
-function buildMessage(
-  diffResult: ReadonlyArray<DiffResult<string>>,
-  { stringDiff = false } = {},
-): string[] {
-  const messages: string[] = [], diffMessages: string[] = [];
-  messages.push("");
-  messages.push("");
-  messages.push(
-    `    ${gray(bold("[Diff]"))} ${red(bold("Actual"))} / ${
-      green(bold("Expected"))
-    }`,
-  );
-  messages.push("");
-  messages.push("");
-  diffResult.forEach((result: DiffResult<string>): void => {
-    const c = createColor(result.type);
-    const line = result.details?.map((detail) =>
-      detail.type !== DiffType.common
-        ? createColor(detail.type, { background: true })(detail.value)
-        : detail.value
-    ).join("") ?? result.value;
-    diffMessages.push(c(`${
-      createSign(result.type)
-    }${line}`));
-  });
-  messages.push(...(stringDiff ? [diffMessages.join("")] : diffMessages));
-  messages.push("");
-
-  return messages;
 }
 
 function isKeyedCollection(x: unknown): x is Set<unknown> {
@@ -255,8 +160,8 @@ export function assertEquals(
     return;
   }
   let message = "";
-  const actualString = _format(actual);
-  const expectedString = _format(expected);
+  const actualString = format(actual);
+  const expectedString = format(expected);
   try {
     const stringDiff = (typeof actual === "string") &&
       (typeof expected === "string");
@@ -342,8 +247,8 @@ export function assertStrictEquals<T>(
   if (msg) {
     message = msg;
   } else {
-    const actualString = _format(actual);
-    const expectedString = _format(expected);
+    const actualString = format(actual);
+    const expectedString = format(expected);
 
     if (actualString === expectedString) {
       const withOffset = actualString
@@ -402,7 +307,7 @@ export function assertNotStrictEquals(
   }
 
   throw new AssertionError(
-    msg ?? `Expected "actual" to be strictly unequal to: ${_format(actual)}\n`,
+    msg ?? `Expected "actual" to be strictly unequal to: ${format(actual)}\n`,
   );
 }
 
@@ -563,9 +468,9 @@ export function assertArrayIncludes(
     return;
   }
   if (!msg) {
-    msg = `actual: "${_format(actual)}" expected to include: "${
-      _format(expected)
-    }"\nmissing: ${_format(missing)}`;
+    msg = `actual: "${format(actual)}" expected to include: "${
+      format(expected)
+    }"\nmissing: ${format(missing)}`;
   }
   throw new AssertionError(msg);
 }
@@ -680,201 +585,6 @@ export function assertObjectMatch(
     // see https://github.com/denoland/deno_std/pull/1419
     filter(expected, expected),
   );
-}
-
-type AssertSnapshotContext = {
-  snapshotPath: string | null;
-  teardownRegistered: boolean;
-  currentSnapshot: Map<string, string> | null;
-  updatedSnapshot: Map<string, string>;
-  snapshotCounts: Map<string, number>;
-  snapshotUpdatedCount: number;
-};
-
-let _assertSnapshotContext: AssertSnapshotContext;
-
-/**
- * Write updates to the snapshot file.
- *
- * @param context Assert snapshot context
- */
-function writeSnapshotFileSync(context: AssertSnapshotContext) {
-  const buf = ["export const snapshot = {};"];
-  function escapeStringForJs(str: string) {
-    return str
-      .replace(/\\/g, "\\\\")
-      .replace(/`/g, "\\`")
-      .replace(/\$/g, "\\$");
-  }
-  context.updatedSnapshot.forEach((value, key) => {
-    let formattedValue = escapeStringForJs(value);
-    formattedValue = formattedValue.includes("\n")
-      ? `\n${formattedValue}\n`
-      : formattedValue;
-    const formattedKey = escapeStringForJs(key);
-    buf.push(`\nsnapshot[\`${formattedKey}\`] = \`${formattedValue}\`;`);
-  });
-  ensureFileSync(context.snapshotPath as string);
-  Deno.writeTextFileSync(context.snapshotPath as string, buf.join("\n"));
-}
-
-/**
- * Register a teardown function which writes the snapshot file to disk and logs the number
- * of snapshots updated after all tests have run.
- *
- * This function can safely be called more than once and will only register the teardown
- * function once.
- *
- * @param context Assert snapshot context
- */
-function registerSnapshotTeardown(context: AssertSnapshotContext) {
-  if (context.teardownRegistered) return;
-  globalThis.onunload = () => {
-    writeSnapshotFileSync(context);
-    if (context.snapshotUpdatedCount > 0) {
-      console.log(
-        green(bold(`\n > ${context.snapshotUpdatedCount} snapshots updated.`)),
-      );
-    }
-  };
-  context.teardownRegistered = true;
-}
-
-/**
- * Make an assertion that `actual` matches a snapshot. If the snapshot and `actual` do
- * not a match, then throw.
- *
- * Type parameter can be specified to ensure values under comparison have the same type.
- * For example:
- * ```ts
- * import { assertSnapshot } from "./asserts.ts";
- *
- * Deno.test("snapshot", async (test) => {
- *  await assertSnapshot<number>(test, 2);
- * });
- * ```
- */
-export async function assertSnapshot(
-  context: Deno.TestContext,
-  actual: unknown,
-  msg?: string,
-): Promise<void>;
-export async function assertSnapshot<T>(
-  context: Deno.TestContext,
-  actual: T,
-  msg?: string,
-): Promise<void>;
-export async function assertSnapshot(
-  context: Deno.TestContext,
-  actual: unknown,
-  msg?: string,
-): Promise<void> {
-  const assertSnapshotContext = getAssertSnapshotContext();
-  const testName = getTestName(context);
-  const count = getCount();
-  const snapshotName = `${testName} ${count}`;
-  const snapshotPath = getSnapshotPath();
-  const isUpdate = Deno.args.some((arg) => arg === "--update" || arg === "-u");
-  const currentSnapshot = await readSnapshotFile();
-
-  const _actual = _format(actual);
-  const _expected = getExpected();
-  if (isUpdate) {
-    if (!equal(_actual, _expected)) {
-      assertSnapshotContext.snapshotUpdatedCount++;
-    }
-    assertSnapshotContext.updatedSnapshot.set(snapshotName, _actual);
-    registerSnapshotTeardown(assertSnapshotContext);
-  } else {
-    if (!_expected) {
-      throw new AssertionError(`Missing snapshot: ${snapshotName}`);
-    }
-    if (equal(_actual, _expected)) {
-      return;
-    }
-    let message = "";
-    try {
-      const stringDiff = (!_actual.includes("\n")) &&
-        (!_actual.includes("\n"));
-      const diffResult = stringDiff
-        ? diffstr(_actual, _expected)
-        : diff(_actual.split("\n"), _expected.split("\n"));
-      const diffMsg = buildMessage(diffResult, { stringDiff }).join("\n");
-      message = `Snapshot does not match:\n${diffMsg}`;
-    } catch {
-      message = `Snapshot does not match:\n${red(CAN_NOT_DISPLAY)} \n\n`;
-    }
-    if (msg) {
-      message = msg;
-    }
-    throw new AssertionError(message);
-  }
-
-  function getAssertSnapshotContext() {
-    if (!_assertSnapshotContext) {
-      _assertSnapshotContext = {
-        snapshotPath: null,
-        teardownRegistered: false,
-        currentSnapshot: null,
-        updatedSnapshot: new Map(),
-        snapshotCounts: new Map(),
-        snapshotUpdatedCount: 0,
-      };
-    }
-    return _assertSnapshotContext;
-  }
-  function getTestName(context: Deno.TestContext): string {
-    if (context.parent) {
-      return `${getTestName(context.parent)} > ${context.name}`;
-    }
-    return context.name;
-  }
-  function getCount() {
-    const count = assertSnapshotContext.snapshotCounts.get(testName) || 1;
-    assertSnapshotContext.snapshotCounts.set(testName, count + 1);
-    return count;
-  }
-  function getSnapshotPath() {
-    if (assertSnapshotContext.snapshotPath) {
-      return assertSnapshotContext.snapshotPath;
-    }
-    const testFile = fromFileUrl(context.origin);
-    const parts = parse(testFile);
-    assertSnapshotContext.snapshotPath = `${
-      join(parts.dir, SNAPSHOT_DIR, parts.name)
-    }${parts.ext}.snap`;
-    return assertSnapshotContext.snapshotPath;
-  }
-  function getExpected() {
-    const snapshot = currentSnapshot.get(snapshotName);
-    if (typeof snapshot === "undefined") {
-      return;
-    }
-    return snapshot.includes("\n") ? snapshot.slice(1, -1) : snapshot;
-  }
-  async function readSnapshotFile() {
-    if (assertSnapshotContext.currentSnapshot) {
-      return assertSnapshotContext.currentSnapshot;
-    }
-    if (isUpdate) {
-      await ensureFile(snapshotPath);
-    }
-    try {
-      const { snapshot } = await import(snapshotPath);
-      assertSnapshotContext.currentSnapshot = typeof snapshot === "undefined"
-        ? new Map()
-        : new Map(Object.entries(snapshot));
-      return assertSnapshotContext.currentSnapshot;
-    } catch (error) {
-      if (
-        error instanceof TypeError &&
-        error.message.startsWith("Module not found")
-      ) {
-        throw new AssertionError("Missing snapshot file.");
-      }
-      throw error;
-    }
-  }
 }
 
 /**
